keeping up untill now changed so many things mostly service now doing the stml

+++ src/main/java/bd/edu/seu/lms/service/BookService.java
@@ -1,6 +1,8 @@
 package bd.edu.seu.lms.service;
 
 import bd.edu.seu.lms.model.Book;
+import bd.edu.seu.lms.model.Allotment;
+import bd.edu.seu.lms.repository.AllotmentRepo;
 import bd.edu.seu.lms.repository.BookRepo;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,14 +13,16 @@ import java.util.stream.Collectors;
 @Service
 public class BookService {
     private final BookRepo bookRepo;
+    private final AllotmentRepo allotmentRepo;
 
-    public BookService(BookRepo bookRepo) {
+    public BookService(BookRepo bookRepo, AllotmentRepo allotmentRepo) {
         this.bookRepo = bookRepo;
+        this.allotmentRepo = allotmentRepo;
     }
 
     @Transactional
     public Book saveBook(Book book) {
-        if (bookRepo.existsById(book.getId())) {
+        if (book.getId() != null && bookRepo.existsById(book.getId())) {
             throw new IllegalArgumentException("Book already exists");
         }
         return bookRepo.save(book);
@@ -34,11 +38,23 @@ public class BookService {
 
     @Transactional
     public void deleteBook(int id) {
-        try {
-            bookRepo.deleteById(id);
-        } catch (Exception e) {
+        if (!bookRepo.existsById(id)) {
             throw new IllegalArgumentException("Book does not exist");
         }
+
+        
+        List<Allotment> allotments = allotmentRepo.findByBook_Id(id);
+
+        if (!allotments.isEmpty()) {
+            String middle="";
+            for(Allotment a:allotments){middle+=(a.getId()+", ");}
+            String s = "Cannot delete book " +
+                    middle
+                    + " allotments of this book still active.";
+            throw new IllegalArgumentException(s);
+        }
+
+        bookRepo.deleteById(id);
     }
 
     public List<Book> getAllBooks() {
