+++ src/main/java/bd/edu/seu/lms/service/VendorService.java
@@ -1,6 +1,8 @@
 package bd.edu.seu.lms.service;
 
+import bd.edu.seu.lms.model.Book;
 import bd.edu.seu.lms.model.Vendor;
+import bd.edu.seu.lms.repository.BookRepo;
 import bd.edu.seu.lms.repository.VendorRepo;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -11,13 +13,17 @@ import java.util.stream.Collectors;
 @Service
 public class VendorService {
     private final VendorRepo vendorRepo;
+    private final BookRepo bookRepo;
 
-    public VendorService(VendorRepo vendorRepo) {
+    public VendorService(VendorRepo vendorRepo, BookRepo bookRepo) {
         this.vendorRepo = vendorRepo;
+        this.bookRepo = bookRepo;
     }
+
     @Transactional
     public Vendor saveVendor(Vendor vendor) {
-        if (vendorRepo.existsById(vendor.getId())) {
+
+        if (vendor.getId() != null && vendorRepo.existsById(vendor.getId())) {
             throw new IllegalArgumentException("Vendor already exist");
         }
         return vendorRepo.save(vendor);
@@ -33,11 +39,24 @@ public class VendorService {
 
     @Transactional
     public void deleteVendor(int id) {
-        try {
-            vendorRepo.deleteById(id);
-        } catch (Exception e) {
+        if (!vendorRepo.existsById(id)) {
             throw new IllegalArgumentException("Vendor does not exist");
         }
+
+        List<Book> books = bookRepo.findByVendor_Id(id);
+
+        if (!books.isEmpty()) {
+            String middle = "";
+            for (Book b : books) {
+                middle += (b.getTitle() + ", ");
+            }
+            String s = "Cannot delete vendor " +
+                    middle
+                    + " books assigned to this vendor.";
+            throw new IllegalArgumentException(s);
+        }
+
+        vendorRepo.deleteById(id);
     }
 
     public List<Vendor> getAllVendors() {
@@ -56,13 +75,11 @@ public class VendorService {
         List<Vendor> nameMatches = vendorRepo.findByNameContainingIgnoreCase(keyword);
         List<Vendor> contactPersonMatches = vendorRepo.findByContactPersonContainingIgnoreCase(keyword);
         List<Vendor> emailMatches = vendorRepo.findByEmailContainingIgnoreCase(keyword);
-        List<Vendor> phoneMatches = vendorRepo.findByPhoneContainingIgnoreCase(keyword);
 
         List<Vendor> combined = new ArrayList<>();
         combined.addAll(nameMatches);
         combined.addAll(contactPersonMatches);
         combined.addAll(emailMatches);
-        combined.addAll(phoneMatches);
         return combined.stream()
                 .distinct()
                 .collect(Collectors.toList());
